"Device's Settings" instead of "DCRS's Settings"
DCRS need to be used to config device, and an operating system for do
works and config all the work ecosystem. But I believe that DCRS and OS
are incorrect names for each one. I'm done for an future when DCRS will
be called as an OS, the operating systems will be called as WS (Work
Systems) and him desktops will be called as WA (Work Ambients). Don't be
chocked by stranger names, people. Get ready for the revolution, get
ready for the better!